requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django
django-cors-headers==4.3.1
djangorestframework==3.15.1
dramatiq==1.16.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
prometheus_client==0.20.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==4.0.0
watchdog-gevent==0.1.1
zope.event==5.0
zope.interface==6.2
